requirements.txt
fastapi
python-telegram-bot==20.6
uvicorn
